Major refactor Minor follow-up cleanup
