+++ emulator/src/gpu/lcd-status-mode.enum.ts
